moving image assets to a separate folder
