+++ src/assets/image_assets/mountain.jpg
