+++ src/assets/image_assets/x_logo.png
