+++ src/components/feature_highlights_section.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import styled from 'styled-components';
 import styles from '../util/styles';
 import { Highlight } from '../util/interface';
-import MountainBackground from '../util/image_assets/mountain.jpg';
-import FitnessBackground from '../util/image_assets/fitness.jpg';
-import TilesBackground from '../util/image_assets/tiles.jpg';
+import MountainBackground from '../assets/image_assets/mountain.jpg';
+import FitnessBackground from '../assets/image_assets/fitness.jpg';
+import TilesBackground from '../assets/image_assets/tiles.jpg';
 import SectionHeaders from './section_headers';
 
 const highlights: Highlight[] = [
